sanjaya_auth_v3/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
cryptography==43.0.3
python-multipart==0.0.12
